Validate table parameters are not empty.
osrm-backend will segfault (on Linux but not OS X) if it is given a
table request with empty parameters. Such a request is non-sensical, so
validate before calling OSRM and return an error.

It would have been nice to return an empty `TableResponse` instead of
returning an error, however it's awkward to do that given it's a thin
wrapper on top of the returned handle. We probably want to know we're
doing empty requests in any case.